+++ NewLayer.py
@@ -1,12 +1,12 @@
 import json
 import requests
 
-req = requests.post("http://localhost:8888/api/v1/layer")
+req = requests.post("http://localhost:8888/api/v1/layer", params={"apikey": "QNCvJkjacXBH"})
 res = json.loads(req.json())
 ds = res["datasource"]
 print(res)
 
-req = requests.get("http://localhost:8888/api/v1/layer/" + ds)
+req = requests.get("http://localhost:8888/api/v1/layer/" + ds, params={"apikey": "QNCvJkjacXBH"})
 res = req.json()
 print(res)
 
@@ -21,15 +21,15 @@ payload = {
 }
 
 print("POST FEATURE")
-req = requests.post("http://localhost:8888/api/v1/layer/" + ds + "/feature", data=json.dumps(payload))
+req = requests.post("http://localhost:8888/api/v1/layer/" + ds + "/feature", params={"apikey": "QNCvJkjacXBH"}, data=json.dumps(payload))
 print(req.json())
 
-print()
-print("GET FEATURE")
-req = requests.get("http://localhost:8888/api/v1/layer/" + ds + "/feature/0")
-print(req.json())
+# print()
+# print("GET FEATURE")
+# req = requests.get("http://localhost:8888/api/v1/layer/" + ds + "/feature/0", params={"apikey": "QNCvJkjacXBH"})
+# print(req.json())
 
-print()
-print("DELETE FEATURE")
-req = requests.delete("http://localhost:8888/api/v1/layer/" + ds)
-print(req.json())
+# print()
+# print("DELETE FEATURE")
+# req = requests.delete("http://localhost:8888/api/v1/layer/" + ds, params={"apikey": "QNCvJkjacXBH"})
+# print(req.json())